// src/app/shared/model/MesaEntrega/PorcentajeMesaModel.ts
export class PorcentajeMesaModel {
    id: number = 0
    catMesaEntregaId: number = null;
    catSemanaId: number = null;
    porcentajeTotalEntrega: number = 0;
    porcentajeTotalValidacion: number = 0;
    porcentajeGeneral: number = 0;
    fechaCreacion: Date = new Date;
    activo: number = 1;
}


  